+++ src/main/java/de/stklcode/jvault/connector/model/response/embedded/SecretMetadata.java
@@ -21,7 +21,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 import java.io.Serializable;
 import java.time.ZonedDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
@@ -36,9 +35,6 @@ import java.util.Objects;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public final class SecretMetadata implements Serializable {
     private static final long serialVersionUID = -905059942871916214L;
-
-    private static final DateTimeFormatter TIME_FORMAT =
-            DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSSSSSSSSXXX");
 
     @JsonProperty("created_time")
     private ZonedDateTime createdTime;
@@ -68,19 +64,6 @@ public final class SecretMetadata implements Serializable {
     private String deleteVersionAfter;
 
     /**
-     * @return Time of secret creation as raw string representation.
-     * @deprecated Method left for backwards compatibility only. Use {@link #getCreatedTime()} instead.
-     */
-    @Deprecated(since = "1.2", forRemoval = true)
-    public String getCreatedTimeString() {
-        if (createdTime != null) {
-            return TIME_FORMAT.format(createdTime);
-        }
-
-        return null;
-    }
-
-    /**
      * @return Time of secret creation.
      */
     public ZonedDateTime getCreatedTime() {
@@ -106,19 +89,6 @@ public final class SecretMetadata implements Serializable {
      */
     public Integer getOldestVersion() {
         return oldestVersion;
-    }
-
-    /**
-     * @return Time of secret update as raw string representation.
-     * @deprecated Method left for backwards compatibility only. Use {@link #getUpdatedTime()} instead.
-     */
-    @Deprecated(since = "1.2", forRemoval = true)
-    public String getUpdatedTimeString() {
-        if (updatedTime != null) {
-            return TIME_FORMAT.format(updatedTime);
-        }
-
-        return null;
     }
 
     /**